requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
